+++ src/Dibi/Drivers/Dummy/Result.php
@@ -7,14 +7,15 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\Dummy;
 
+use Dibi\Drivers;
 
 
 /**
  * The driver for no result set.
  */
-class NoDataResult implements Result
+class Result implements Drivers\Result
 {
 	public function __construct(
 		private readonly int $rows,
