+++ src/Dibi/Drivers/Engines/MySQLEngine.php
@@ -7,16 +7,18 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\Engines;
 
 use Dibi;
+use Dibi\Drivers\Connection;
+use Dibi\Drivers\Engine;
 
 
 /**
  * The reflector for MySQL databases.
  * @internal
  */
-class MySqlReflector implements Engine
+class MySQLEngine implements Engine
 {
 	public function __construct(
 		private readonly Connection $driver,
