+++ src/Dibi/Drivers/Engines/PostgreSQLEngine.php
@@ -7,14 +7,16 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\Engines;
 
+use Dibi\Drivers\Connection;
+use Dibi\Drivers\Engine;
 
 
 /**
  * The reflector for PostgreSQL database.
  */
-class PostgreReflector implements Engine
+class PostgreSQLEngine implements Engine
 {
 	public function __construct(
 		private readonly Connection $driver,
