+++ src/Dibi/Drivers/Firebird/Connection.php
@@ -7,9 +7,10 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\Ibase;
 
 use Dibi;
+use Dibi\Drivers;
 use Dibi\Helpers;
 
 
@@ -24,7 +25,7 @@ use Dibi\Helpers;
  *   - buffers (int) => buffers is the number of database buffers to allocate for the server-side cache. If 0 or omitted, server chooses its own default.
  *   - resource (resource) => existing connection resource
  */
-class FirebirdDriver implements Connection
+class Connection implements Drivers\Connection
 {
 	public const ErrorExceptionThrown = -836;
 
@@ -199,9 +200,9 @@ class FirebirdDriver implements Connection
 	/**
 	 * Returns the connection reflector.
 	 */
-	public function getReflector(): Engine
+	public function getReflector(): Drivers\Engine
 	{
-		return new FirebirdReflector($this);
+		return new Drivers\Engines\FirebirdEngine($this);
 	}
 
 
@@ -209,9 +210,9 @@ class FirebirdDriver implements Connection
 	 * Result set driver factory.
 	 * @param  resource  $resource
 	 */
-	public function createResultDriver($resource): FirebirdResult
+	public function createResultDriver($resource): Result
 	{
-		return new FirebirdResult($resource);
+		return new Result($resource);
 	}
